add bundle splitting
