+++ app/index.tsx
@@ -1,90 +1,19 @@
-import MuiThemeProvider from "material-ui/styles/MuiThemeProvider";
 import * as React from "react";
 import * as ReactDOM from "react-dom";
-import { Provider } from "react-redux";
-import { applyMiddleware, compose, createStore } from "redux";
-import { composeWithDevTools } from "redux-devtools-extension";
-import reduxLogger from "redux-logger";
-import { autoRehydrate, persistStore } from "redux-persist";
-import reduxThunk from "redux-thunk";
+import { Provider, Store } from "react-redux";
 
-import { asyncSessionStorage } from "redux-persist/storages";
 import App from "./containers/App";
-import Commit from "./containers/Commit";
-import muiTheme from "./muiTheme";
-import reducers from "./reducers";
-import { initRepository } from "./web3/contracts/ContractsRepository";
+import { startup } from "./startup";
 
-// @todo add bundle splitting and separate these renders #UAF
-
-const render = (storage: any) => {
-  /* We are doing this because we are not loading the "react-root"
-  div in the following pages[whitepaper, faq, product]
-  */
+const render = (storage: Store<any>) => {
   const indexRoot = document.getElementById("react-root");
-  const commitRoot = document.getElementById("react-root-commit");
-  if (indexRoot) {
-    ReactDOM.render(
-      <Provider store={storage}>
-        <App />
-      </Provider>,
-      indexRoot
-    );
-  }
-  if (commitRoot) {
-    ReactDOM.render(
-      <MuiThemeProvider muiTheme={muiTheme}>
-        <Provider store={storage}>
-          <Commit />
-        </Provider>
-      </MuiThemeProvider>,
-      commitRoot
-    );
-  }
+
+  ReactDOM.render(
+    <Provider store={storage}>
+      <App />
+    </Provider>,
+    indexRoot
+  );
 };
 
-const enhancers = () =>
-  composeWithDevTools(compose(applyMiddleware(reduxThunk, reduxLogger), autoRehydrate()));
-
-// Create the Redux store
-const store = createStore(reducers, enhancers());
-
-// Add development time features
-if (process.env.NODE_ENV !== "production") {
-  // Enable React Debug Tool
-  // tslint:disable-next-line
-  const ReactDebugTool = require('react-dom/lib/ReactDebugTool');
-  ReactDebugTool.beginProfiling();
-
-  // Enable Webpack hot module replacement
-  if ((module as any).hot) {
-    // Replace reducers
-    // (module as any).hot.accept('./reducers', () => {
-    //   const newReducer = require('./reducers').default;
-    //   store.replaceReducer(newReducer);
-    // });
-    // Replace routes and components
-    // (module as any).hot.accept('./routes', () => {
-    //   const newAppRoutes = require('./routes').default;
-    //   // NOTE: We don't update the store middleWare,
-    //   //       if you change routes you need to refresh.
-    //   render(store, newAppRoutes);
-    // });
-  }
-
-  // we require this files only to track changes in them automatically
-
-  // tslint:disable-next-line
-  require('!raw-loader!../dist/index.html');
-  // tslint:disable-next-line
-  require('!raw-loader!../dist/app.css');
-}
-
-persistStore(
-  store,
-  {
-    whitelist: ["legalAgreementState"],
-    storage: asyncSessionStorage,
-  },
-  () => initRepository().then(() => render(store))
-);
+startup(render);
