+++ page/commit.ejs
@@ -4,7 +4,8 @@
     <body>
         <div id="react-root-commit"></div>
 
-        <script src="/main<%= hashes.jsBundleHash %>.js"></script>
+        <script src="/common<%= hashes.jsBundleHash %>.js"></script>
+        <script src="/commit<%= hashes.jsBundleHash %>.js"></script>
         <noscript>
           <p>This page requires JavaScript to be enabled.</p>
         </noscript>
