+++ page/index.ejs
@@ -9,7 +9,9 @@
         <div>
           <%- include('partials/home'); %>
 
+          <script src="/common<%= hashes.jsBundleHash %>.js"></script>
           <script src="/main<%= hashes.jsBundleHash %>.js"></script>
+          <script src="/page<%= hashes.jsBundleHash %>.js"></script>
           <noscript>
             <p>This page requires JavaScript to be enabled.</p>
           </noscript>
