+++ page/product.ejs
@@ -19,7 +19,7 @@
           <% } %>
           <%- include('partials/footer', content.footer); %>
 
-          <script src="/main<%= hashes.jsBundleHash %>.js"></script>
+          <script src="/page<%= hashes.jsBundleHash %>.js"></script>
           <noscript>
             <p>This page requires JavaScript to be enabled.</p>
           </noscript>
